requirements.txt
mcp[cli]
anthropic
gitpython
google-api-python-client
google-auth
python-dotenv
pandas
openpyxl